fully implement strategy config and work on contract structure
CU-1yf802z CU-20r9088 CU-1yf7vry

+++ contracts/interfaces/IBoostPool.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.0;
 
 interface IBoostPool {
-    function bonusEndBlock() external view returns (uint256);
+    function bonusEndBlock() external view returns (uint64);
     function STAKE_TOKEN_VID() external view returns (uint256);
-    function vaultHealerActivate(uint _boostID) external;
+    function vaultHealerActivate(uint64 _boostID) external;
     function joinPool(address _user, uint _amount) external;
     function harvest(address) external;
     function emergencyWithdraw(address _user) external returns (bool success);
